+++ src/actions/login.js
@@ -1,8 +1,0 @@
-import { AUTH_REQUEST } from '../actions/actionTypes'
-
-const loginRequest = payload => ({
-  type: AUTH_REQUEST,
-  payload
-})
-
-export default loginRequest