+++ src/reducers/me.js
@@ -1,10 +1,19 @@
-import { AUTH_SUCCESS } from '../actions/actionTypes'
+import { AUTH_SUCCESS, LOGOUT_REQUEST } from '../actions/actionTypes'
 
 const initialState = {
   data: {
     auth: false,
-    fullName: 'John Doe',
-    email: 'john@doe.com'
+    full_name: 'J',
+    email: '',
+    client: {
+      full_name: '',
+      phone: '',
+      address_1: '',
+      address_2: '',
+      city: '',
+      state: '',
+      zip: ''
+    }
   }
 }
 
@@ -15,6 +24,9 @@ const reducer = (state = initialState, action) => {
         ...state,
         data: action.payload
       }
+
+    case LOGOUT_REQUEST:
+      return initialState
 
     default:
       return state